+++ tests/unit-tests/boomerang/ssl/type/TypeTest.cpp
@@ -115,10 +115,12 @@ void TypeTest::testCompound()
     prog->readDefaultLibraryCatalogues();
 
     std::shared_ptr<Signature> paintSig = prog->getLibSignature("BeginPaint");
-
     SharedType paramType = paintSig->getParamType(1);
+
+    QVERIFY(paramType != nullptr);
     QCOMPARE(paintSig->getParamType(1)->getCtype(), QString("LPPAINTSTRUCT"));
     SharedType paintStructType = paramType->as<PointerType>()->getPointsTo();
+    QVERIFY(paintStructType != nullptr);
     QCOMPARE(paintStructType->getCtype(), QString("PAINTSTRUCT"));
 
     // Offset 8 should have a RECT